libs/requirements.txt
nose==1.3.4
pycrypto==2.6.1
wsgiref==0.1.2
requests==2.6.0
beautifulsoup4==4.4.1
lxml
oss2
requests
django-redis==4.4.4